+++ deepcave/plugins/__init__.py
@@ -115,9 +115,9 @@ class Plugin(Layout, ABC):
         str
             Url for the plugin as string.
         """
-        from deepcave import config
+        from deepcave.config import Config
 
-        return f"http://{config.DASH_ADDRESS}:{config.DASH_PORT}/plugins/{cls.id}"
+        return f"http://{Config.DASH_ADDRESS}:{Config.DASH_PORT}/plugins/{cls.id}"
 
     @staticmethod
     def check_run_compatibility(run: AbstractRun) -> bool:
@@ -395,22 +395,28 @@ class Plugin(Layout, ABC):
                             # because `run` would be removed.
                             # Also: The current run name does not need to be kept.
                             update_dict(_inputs, self.load_inputs())
+                            # Reset inputs
+                            if "objective_id" in _inputs.keys():
+                                update_dict(_inputs, {"objective_id": {"value": None}})
+                            if "budget_id" in _inputs.keys():
+                                update_dict(_inputs, {"budget_id": {"value": None}})
+                            if "hyperparameter_name_1" in _inputs.keys():
+                                update_dict(_inputs, {"hyperparameter_name_1": {"value": None}})
+                            if "hyperparameter_name_2" in _inputs.keys():
+                                update_dict(_inputs, {"hyperparameter_name_2": {"value": None}})
 
-                            # TODO: Reset only inputs which are not available in another run.
-                            # E.g. if options from budget in run_2 and run_3 are the same
-                            # take the budget from run_2 if changed to run_3. Otherwise,
-                            # reset budgets.
+                        if _run_id:
+                            selected_run = run_handler.get_run(_run_id)
 
-                        selected_run = run_handler.get_run(_run_id)
+                    if selected_run is not None:
+                        # How to update only parameters which have a dependency?
+                        user_dependencies_inputs = self.load_dependency_inputs(
+                            selected_run, _previous_inputs, _inputs
+                        )
 
-                    # How to update only parameters which have a dependency?
-                    user_dependencies_inputs = self.load_dependency_inputs(
-                        selected_run, _previous_inputs, _inputs
-                    )
-
-                    # Update dict
-                    # dict.update() removes keys, so our own method is used to do so
-                    update_dict(inputs, user_dependencies_inputs)  # inplace operation
+                        # Update dict
+                        # dict.update() removes keys, so our own method is used to do so
+                        update_dict(inputs, user_dependencies_inputs)  # inplace operation
 
                 # Let's cast the inputs
                 inputs = self._cast_inputs(inputs)