+++ examples/api/pdp.py
@@ -1,12 +1,12 @@
 """
-Parallel Coordinates
+Partial Dependencies
 ^^^^^^^^^^^^^^^^^^^^
 
-This example shows how to use the plugin Parallel Coordinates.
+This example shows how to use the plugin Partial Dependencies.
 Note that other plugins use the same interfaces and can be used in the same fashion.
 """
 
-from deepcave.plugins.objective.parallel_coordinates import ParallelCoordinates
+from deepcave.plugins.hyperparameter.pdp import PartialDependencies
 from deepcave.runs.converters.deepcave import DeepCAVERun
 from pathlib import Path
 
@@ -18,15 +18,14 @@ if __name__ == "__main__":
     budget_id = run.get_budget_ids()[-1]
 
     # Instantiate the plugin
-    plugin = ParallelCoordinates()
+    plugin = PartialDependencies()
     inputs = plugin.generate_inputs(
-        hyperparameter_names=run.configspace.get_hyperparameter_names(),
+        hyperparameter_name_1="alpha",
+        hyperparameter_name_2="beta",
         objective_id=objective_id,
         budget_id=budget_id,
-        show_important_only=True,
-        show_unsuccessful=False,
-        hide_hps=False,
-        n_hps=3,
+        show_confidence=False,
+        show_ice=True,
     )
     # Note: Filter variables are not considered.
     outputs = plugin.generate_outputs(run, inputs)
@@ -34,5 +33,5 @@ if __name__ == "__main__":
     # Finally, you can load the figure. Here, the filter variables play a role.
     # Alternatively: Use the matplotlib output (`load_mpl_outputs`) if available.
     figure = plugin.load_outputs(run, inputs, outputs)  # plotly.go figure
-    figure.write_image("examples/api/parallel_coordinates.png", scale=2.0)
+    figure.write_image("examples/api/pdp.png")
     # figure.show()