+++ tests/test_utils/test_cache.py
@@ -302,7 +302,7 @@ class TestLogger(unittest.TestCase):
         self.assertFalse(mpl_logger.propagate)
 
         plugin_logger = get_logger("src.plugins")
-        self.assertEqual(logging.DEBUG, plugin_logger.level)
+        self.assertEqual(logging.INFO, plugin_logger.level)
         self.assertFalse(plugin_logger.propagate)